Big refactor to make sense of code
Use sensible names, and also fix a few bug discovered by altering the
tests with more data and cases (such as the SQLDelete statement). CVE
DataObject is now SecurityAlerts, as not all alerts are assigned CVE
identifiers. To that end if there is another identifier (such as the one
SilverStripe use as a title prefix) then this is set instead. As such the
property is renamed 'Identifier' as opposed to 'CVE'.